+++ src/Bauelemente/Widerstand.java
@@ -3,12 +3,9 @@ package Bauelemente;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 
-/**
- * Created by Merve on 24.04.2017.
- */
 public class Widerstand extends Bauelement {
 
-    public Widerstand(int x, int y, int Orientation)
+    public Widerstand(double x, double y, int Orientation)
     {
         super(x,y,Orientation);
         System.out.println("Class Wid: "+posX+","+posY+","+Orientation);
@@ -17,4 +14,5 @@ public class Widerstand extends Bauelement {
     {
         gc.drawImage(image,posX-25,posY-25);
     }
+
 }
